Fix salt-api not starting and ensure the pip module installs into the desired relenv directory by default
